pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robotframework-agent"
version = "0.0.1"
description = "AI-driven test automation agent for apps"
readme = "readme.md"
requires-python = ">=3.10"
license = {text = "CC0-1.0"}
authors = [
    {name = "Abdelkader HASSINE", email = "contact.abdelkaderhassine@gmail.com"},
    {name = "Pavlo Ivashchenko", email = "pawloiwanov@gmail.com"},
]
keywords = ["robotframework", "ai", "automation", "automation agent", "llm", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "License :: CC0 1.0 Universal (CC0 1.0) Public Domain Dedication",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Robot Framework",
    "Framework :: Robot Framework :: Library",
]

dependencies = [
    "robotframework>=6.0",
    "openai>=1.0.0",
    "Pillow>=10.0.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
web = [
    "playwright>=1.40.0",
]
mobile = [
    "robotframework-appiumlibrary>=3.0",
    "Appium-Python-Client>=5.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]
dev = [
    "playwright>=1.40.0",
    "robotframework-appiumlibrary>=3.0",
    "Appium-Python-Client>=5.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.0",
    "pre-commit>=3.6.0",
    "anthropic>=0.20.0",
    "google-generativeai>=0.3.0",
    "gradio-client>=2.0.0",
]
browserstack = [
    "browserstack-sdk>=1.30.0",
    "browserstack-local>=1.2.0",
]
anthropic = [
    "anthropic>=0.20.0",
]
gemini = [
    "google-generativeai>=0.3.0",
]
vision = [
    "gradio-client>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/aidriventesting"
Documentation = "https://github.com/aidriventesting/agent#readme"
Repository = "https://github.com/aidriventesting/agent"
Issues = "https://github.com/aidriventesting/agent/issues"

[tool.setuptools]
packages = [
    "Agent",
    "Agent.ai",
    "Agent.ai.llm",
    "Agent.ai.vlm",
    "Agent.ai.prompts",
    "Agent.config",
    "Agent.core",
    "Agent.platforms",
    "Agent.platforms.collectors",
    "Agent.platforms.locators",
    "Agent.tools",
    "Agent.tools.mobile",
    "Agent.tools.visual",
    "Agent.tools.web",
    "Agent.utilities",
    "Agent.utilities.imguploader",
]

[tool.setuptools.package-data]
Agent = ["config/*.json"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.pytest.ini_options]
testpaths = ["tests/utest"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

